rename proto package to protocol
