+++ core/pkg/protocol/http.go
@@ -1,4 +1,4 @@
-package proto
+package protocol
 
 import (
 	"fmt"
